Refractor BiomeUtil
