+++ platforms/mixin-common/src/main/java/com/dfsek/terra/mod/util/TagUtil.java
@@ -54,7 +54,7 @@ public final class TagUtil {
         logger.info("Doing biome tag garbage....");
         Map<TagKey<Biome>, List<RegistryEntry<Biome>>> collect = tagsToMutableMap(registry);
 
-        MinecraftUtil
+        BiomeUtil
             .getTerraBiomeMap()
             .forEach((vb, terraBiomes) ->
                 MinecraftUtil
