+++ platforms/mixin-lifecycle/src/main/java/com/dfsek/terra/lifecycle/LifecyclePlatform.java
@@ -22,7 +22,7 @@ import java.util.stream.Stream;
 
 import com.dfsek.terra.addon.EphemeralAddon;
 import com.dfsek.terra.api.addon.BaseAddon;
-import com.dfsek.terra.lifecycle.util.BiomeUtil;
+import com.dfsek.terra.lifecycle.util.LifecycleBiomeUtil;
 import com.dfsek.terra.mod.CommonPlatform;
 import com.dfsek.terra.mod.ModPlatform;
 import com.dfsek.terra.mod.generation.MinecraftChunkGeneratorWrapper;
@@ -68,7 +68,7 @@ public abstract class LifecyclePlatform extends ModPlatform {
 
 
         if(server != null) {
-            BiomeUtil.registerBiomes(server.getRegistryManager().get(RegistryKeys.BIOME));
+            LifecycleBiomeUtil.registerBiomes(server.getRegistryManager().get(RegistryKeys.BIOME));
             server.reloadResources(server.getDataPackManager().getNames()).exceptionally(throwable -> {
                 LOGGER.warn("Failed to execute reload", throwable);
                 return null;
