+++ platforms/mixin-lifecycle/src/main/java/com/dfsek/terra/lifecycle/util/LifecycleUtil.java
@@ -18,7 +18,7 @@ public final class LifecycleUtil {
 
     public static void initialize(MutableRegistry<Biome> biomeMutableRegistry, MutableRegistry<WorldPreset> worldPresetMutableRegistry) {
         CommonPlatform.get().getEventManager().callEvent(new PlatformInitializationEvent());
-        BiomeUtil.registerBiomes(biomeMutableRegistry);
+        LifecycleBiomeUtil.registerBiomes(biomeMutableRegistry);
         CommonPlatform.get().registerWorldTypes(
             (id, preset) -> Registry.register(worldPresetMutableRegistry, RegistryKey.of(RegistryKeys.WORLD_PRESET, id), preset));
     }
